third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-without-allow-attribute-allowed-from-parent-expected.txt

CONSOLE WARNING: Resource requests whose URLs contain raw newline characters are deprecated, and may be blocked in M60, around August 2017. Please remove newlines from places like element attribute values in order to continue loading those resources. See https://www.chromestatus.com/features/5735596811091968 for more details.
Test entering full screen security restrictions. An iframe without an allow attribute is still permitted to fullscreen if the request comes from the containing document.

To test manually, press any key - the page should enter full screen mode.

TEST(document.webkitFullscreenEnabled) OK
iframe's webkitFullscreenEnabled should be false:
TEST(iframeMessage == 'document.webkitFullscreenEnabled == false') OK
EVENT(webkitfullscreenchange)
SUCCEED - entered full screen!
END OF TEST
